+++ Frends.Avro.Serialize/CHANGELOG.md
@@ -1,4 +1,13 @@
-# Changelog
+﻿# Changelog
+
+## [2.0.0] - 2025-08-07
+
+### Changed 
+- [Breaking] Renamed `OutputPath` parameter to `TargetFilePath` in Input tab
+- Updated Result structure to include Success boolean and Error object
+
+### Added 
+- Added new Options tab with error handling parameters (ThrowErrorOnFailure, ErrorMessageOnFailure)
 
 ## [1.0.0] - 2024-05-23
 ### Added